requirements/requirements.txt
PyJWT~=2.9.0
Flask~=3.0.3
python-dotenv~=1.0.1
requests~=2.32.3
xlsxwriter~=3.2.0